Extract and use HttpListenerRequest's helper class for managed implementation
